update: refactor code from EOS to FO
