+++ advanced/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "advanced",
+  "version": "1.0.0",
+  "description": "",
+  "repository": "",
+  "keywords": "",
+  "author": "",
+  "license": "ISC"
+}
