+++ basic/hello_fibos/fibos_client/test/sample_test.js
@@ -1,0 +1,16 @@
+var test = require('test');
+test.setup();
+
+describe('a sample case', () => {
+    var name;
+
+    before(() => {
+        name = "FIBOS";
+    });
+
+    it('check name', () => {
+        assert.equal(name, "FIBOS");
+    });
+});
+
+test.run();
