+++ basic/hello_fibos/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "hello_fibos",
+  "version": "1.0.0",
+  "description": "",
+  "repository": "",
+  "keywords": "",
+  "author": "",
+  "license": "ISC"
+}
