+++ community/fibos_eos/sync_eos.js
