+++ development/testNet/about/client.js
@@ -1,0 +1,6 @@
+var FIBOS = require("./initClient.js");
+var fibos = FIBOS("");
+
+var result = fibos.getBlockSync(1);
+
+console.log(result);
