--- hello_fibos/fibos_client/sample_test.js
@@ -1,16 +1,0 @@
-var test = require('test');
-test.setup();
-
-describe('a sample case', () => {
-	var name;
-
-	before(() => {
-		name = "FIBOS";
-	});
-
-	it('check name', () => {
-		assert.equal(name, "FIBOS");
-	});
-});
-
-test.run();
